adds constant of authorized_keys

+++ rita_client/src/operator_update/mod.rs
@@ -21,7 +21,7 @@ use rita_common::rita_loop::is_gateway;
 use rita_common::tunnel_manager::neighbor_status::get_neighbor_status;
 use rita_common::tunnel_manager::shaping::flag_reset_shaper;
 use rita_common::utils::option_convert;
-use rita_common::DROPBEAR_CONFIG;
+use rita_common::DROPBEAR_AUTHORIZED_KEYS;
 use rita_common::KI;
 use serde_json::Map;
 use serde_json::Value;
@@ -258,7 +258,6 @@ async fn checkin() {
         (_, Some(new)) => Some(new),
     };
     set_router_update_instruction(update_instructions);
-
     perform_operator_update(new_settings, rita_client, network)
 }
 
@@ -337,8 +336,8 @@ fn perform_operator_update(
             add_list,
             drop_list,
         }) => {
-            let key_file = DROPBEAR_CONFIG;
-            info!("Updating {}", key_file);
+            let key_file = DROPBEAR_AUTHORIZED_KEYS;
+            info!("Updating auth_keys {:?}", key_file);
             let res = update_authorized_keys(add_list, drop_list, key_file);
             info!("Update auth_keys result is  {:?}", res);
         }
@@ -361,7 +360,7 @@ fn update_authorized_keys(
     let mut existing = HashSet::new();
     let auth_keys_file = File::open(keys_file);
     let mut write_data: Vec<String> = vec![];
-    let temp_key_file = String::from("key_backup");
+    let temp_key_file = String::from("/etc/dropbear/key_backup");
 
     info!(
         "Authorized keys updates add {} remove {} pubkeys",
@@ -404,6 +403,8 @@ fn update_authorized_keys(
                 managed: true,
                 flush: false,
             });
+        } else {
+            info!("Authorized keys failed to parse key {:?}", pubkey);
         }
     }
     // parse list for keys to remove, setting flush = true
@@ -427,22 +428,26 @@ fn update_authorized_keys(
             write_data.push(key.key.to_string());
         }
     }
+    info!("DEBUG: {:#?}", &existing);
+
     // create string block to use a single write to temp file
-    match write!(&updated_key_file, "{}", write_data.join("\n")) {
-        Ok(()) => info!("Authorized keys write success"),
+    match write!(&updated_key_file, "{}", &write_data.join("\n")) {
+        Ok(()) => info!("Authorized keys write success {:#?}", write_data),
         Err(e) => info!("Authorized keys write failed with {:?}", e),
     };
 
     // rename temp file
-    match rename(&temp_key_file, keys_file) {
-        Ok(()) => {
-            info!("Authorized keys rename success")
-        }
-        Err(e) => {
-            info!("Authorized keys rename failed with {:?}", e);
-            remove_file(&temp_key_file)?
-        }
-    };
+    if updated_key_file.metadata().unwrap().len() != 0 {
+        match rename(&temp_key_file, keys_file) {
+            Ok(()) => {
+                info!("Authorized keys rename success")
+            }
+            Err(e) => {
+                info!("Authorized keys rename failed with {:?}", e);
+                remove_file(&temp_key_file)?
+            }
+        };
+    }
 
     Ok(())
 }
